+++ requirements.txt
@@ -1,3 +1,4 @@
+dnspython==2.7.0
 psycopg==3.2.4
 psycopg2-binary==2.9.10
 python-dotenv==1.0.1
